pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dodo-rl"
version = "0.1.0"
description = "Robot locomotion training pipeline with Isaac Lab"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.5.0",
    "numpy>=1.26",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "wandb>=0.18",
    "tensorboard>=2.18",
    "huggingface-hub>=0.26",
    "onnx>=1.17",
    "onnxruntime>=1.20",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
    "pytest-xdist",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["isaaclab_tasks*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
